+++ src/components/PageNotFound.vue
@@ -1,0 +1,12 @@
+<template lang="html">
+  <div>
+    Error - Page Not Found
+  </div>
+</template>
+<script>
+export default {
+  name: 'PageNotFound'
+}
+</script>
+<style scoped>
+</style>