+++ src/main.js
@@ -1,13 +1,39 @@
 import Vue from 'vue'
 import App from './App.vue'
 import router from './router'
-import store from './store'
+import { store } from './store'
 import { BootstrapVue, BootstrapVueIcons, IconsPlugin } from 'bootstrap-vue'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { fas } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import Vuelidate from 'vuelidate'
+import VueLogger from 'vuejs-logger';
+const isProduction = process.env.NODE_ENV === 'production';
 
+const logOptions = {
+  isEnabled: true,
+  logLevel : isProduction ? 'error' : 'debug',
+  stringifyArguments : false,
+  showLogLevel : true,
+  showMethodName : true,
+  separator: '-',
+  showConsoleColors: true
+};
+
+// global prototypes for input
+Vue.prototype.$questions = {}
+Vue.prototype.$required = {}
+// Collection object of input prototypes
+Vue.prototype.$input_object = {}
+
+// global prototypes for output
+Vue.prototype.$values = {}
+Vue.prototype.$logging_object = {}
+// Collection object of output prototypes
+Vue.prototype.$output_object = {}
+
+// Use simple logger
+Vue.use(VueLogger, logOptions);
 // Install BootstrapVue
 Vue.use(BootstrapVue)
 // Optionally install the BootstrapVue icon components plugin
@@ -17,9 +43,10 @@ Vue.use(IconsPlugin)
 // Optionally install vuelidatejs plugin
 Vue.use(Vuelidate)
 
+//Vue.use(OverlayPlugin)
+
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
-// import './custom.scss'
 
 // This imports all the layout components such as <b-container>, <b-row>, <b-col>:
 import { LayoutPlugin } from 'bootstrap-vue'
@@ -45,10 +72,27 @@ Vue.use(TablePlugin)
 library.add(fas)
 Vue.component('font-awesome-icon', FontAwesomeIcon)
 
+// Custom components
+import Home from '@/components/Home.vue'
+import Daacs from '@/components/Daacs.vue'
+import Questions from '@/components/Questions.vue'
+import Help from '@/components/Help.vue'
+import PageNotFound from '@/components/PageNotFound.vue'
+
+Vue.component('Daacs', Daacs);
+Vue.component('Home', Home);
+Vue.component('PageNotFound', PageNotFound);
+Vue.component('Questions', Questions);
+Vue.component('Help', Help);
+
 Vue.config.productionTip = false
 
 new Vue({
   router,
   store,
+  created(){
+  },
+  mounted(){
+  },
   render: h => h(App)
 }).$mount('#app')